reduce iterations on images and cache opened image for better performance
Instead of doing findIndex multiple times when image is changed to next or previous, use findIndex only when modal which shows the chosen image is opened, than save image index and image id. In changeImageHandler, operations which change image to next or previous are based on the saved index which reduce unnecessary iterations.

When image is opened cache its id and index.

If the modal is opened with the same image consecutively, get its index from cache instead of using findIndex to reduce the unnecessary iterations.